scattering..
