+++ scatterNgather/iosScatter&GatherProject/iosScatter&GatherProject.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "self:iosScatter&amp;GatherProject.xcodeproj">
+      location = "self:">
    </FileRef>
 </Workspace>
