+++ scatterNgather/iosScatter&GatherProject/iosScatter&GatherProject.xcodeproj/project.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
