+++ scatterNgather/iosScatter&GatherProject/iosScatter&GatherProject/AppDelegate.swift
