+++ scatterNgather/iosScatter&GatherProject/iosScatter&GatherProject/Assets.xcassets/AppIcon.appiconset/Contents.json
