+++ scatterNgather/iosScatter&GatherProject/iosScatter&GatherProject/Assets.xcassets/Contents.json
