+++ scatterNgather/iosScatter&GatherProject/iosScatter&GatherProject/Assets.xcassets/lambda_logo.imageset/Contents.json
