+++ scatterNgather/iosScatter&GatherProject/iosScatter&GatherProject/Info.plist
